commit a21f3fad19b6ccb8998f4cc7178bf79b66635c58
Author: Sylvain Doremus <dragonjoker59@hotmail.com>
Date:   Fri Dec 26 18:14:32 2014 +0100

    x64 compatibility on Windows.

+++ source/Core/Castor3D/Src/OverlayRenderer.cpp
@@ -137,7 +137,7 @@ namespace Castor3D
 		if ( l_pMaterial )
 		{
 			std::vector< OverlayCategory::Vertex > const & l_arrayVtx = p_panelOverlay.GetVertexBuffer();
-			size_t l_count = l_arrayVtx.size();
+			uint32_t l_count = uint32_t( l_arrayVtx.size() );
 
 			if ( m_previousPanelZIndex != iZIndex )
 			{
@@ -175,7 +175,7 @@ namespace Castor3D
 		if ( l_pMaterial )
 		{
 			std::vector< OverlayCategory::Vertex > const & l_arrayVtx = p_borderPanelOverlay.GetVertexBuffer();
-			size_t l_count = l_arrayVtx.size();
+			uint32_t l_count = uint32_t( l_arrayVtx.size() );
 
 			if ( m_previousPanelZIndex != iZIndex )
 			{
@@ -209,7 +209,7 @@ namespace Castor3D
 		if ( l_pMaterial )
 		{
 			std::vector< OverlayCategory::Vertex > const & l_arrayVtx = p_borderPanelOverlay.GetBorderVertexBuffer();
-			size_t l_count = l_arrayVtx.size();
+			uint32_t l_count = uint32_t( l_arrayVtx.size() );
 
 			if ( m_previousBorderZIndex != iZIndex )
 			{
@@ -253,7 +253,7 @@ namespace Castor3D
 			if ( l_pMaterial )
 			{
 				std::vector< OverlayCategory::Vertex > const & l_arrayVtx = p_textOverlay.GetVertexBuffer();
-				size_t l_count = l_arrayVtx.size();
+				uint32_t l_count = uint32_t( l_arrayVtx.size() );
 
 				if ( m_previousCaption != p_textOverlay.GetCaption() || m_previousTextZIndex != iZIndex )
 				{
